backend/requirements.txt
fastapi>=0.115
uvicorn>=0.30
networkx>=3.2
langgraph>=0.2
langchain-google-genai>=2.0
google-generativeai>=0.8
pydantic>=2.9
python-dotenv>=1.0
scipy>=1.13
numpy>=2.0
